requirements.txt
Django==1.5.4
ipython==1.1.0
django-braces==1.2.2
django-model-utils==1.5.0
logutils==0.3.3
South==0.8.2
pygithub3==0.5
django-social-auth==0.7.28
django-tastypie==0.10.0
psycopg2==2.5.1
factory-boy==2.1.2
mock==1.0.1
path.py==4.3
coverage==3.6
django-discover-runner==1.0
django-debug-toolbar==0.9.4
Sphinx==1.1.3
django-nose==1.2
nose==1.3.0
gunicorn==18.0
